# Copy validated declarations between checker.Env instances on extension