@TristonianJones 2021-02-18
@JimLarson Do you think you'd be able to take a look at this? I realize more extensive refactors are probably necessary, but looking this change again, it does make the code easier to decipher (imo).