@TristonianJones 2022-03-24
@jpbetz Once the release is cut would you mind just updating the #519 ticket with the updated performance graph?

Closes #346